requirements.txt
python-dotenv>=1.0.0
torch>=2.0.0
numpy>=1.24.0
gymnasium>=0.29.0
transformers>=4.30.0
datasets>=2.14.0
scipy>=1.10.0
scikit-learn>=1.3.0
pandas>=2.0.0
matplotlib>=3.7.0
tensorboard>=2.13.0
jupyterlab>=4.0.0
ipykernel>=6.23.0
